backend/requirements.txt
Flask==3.1.0
Flask-Cors==5.0.1
Flask-Limiter==3.9.4
PyJWT==2.10.1
python-dotenv==1.1.0
supabase==2.28.3
marshmallow==3.26.1
